# Add test to measure end-to-end performance on pinned / non-pinned memory