--- a/Dockerfile.QA
+++ b/Dockerfile.QA
@@ -133,6 +133,8 @@ RUN cp builddir/trtis-custom-backends/install/lib/libidentity.so \
         qa/L0_model_config/. && \
     cp builddir/trtis-custom-backends/install/lib/libidentity.so \
         qa/L0_simple_lib/. && \
+    cp builddir/trtis-custom-backends/install/lib/libidentity.so \
+        qa/L0_pinned_memory/. && \
     mkdir -p qa/L0_infer_shm && \
     cp -r qa/L0_infer/. qa/L0_infer_shm && \
     mkdir -p qa/L0_batcher_shm && \