--- a/qa/L0_pinned_memory/libtorch_ensemble.pbtxt
+++ b/qa/L0_pinned_memory/libtorch_ensemble.pbtxt
@@ -1,0 +1,45 @@
+name: "libtorch_ensemble"
+platform: "ensemble"
+max_batch_size: 0
+input [
+  {
+    name: "INPUT0"
+    data_type: TYPE_FP32
+    dims: [ -1 ]
+  }
+]
+output [
+  {
+    name: "OUTPUT0"
+    data_type: TYPE_FP32
+    dims: [ -1 ]
+  }
+]
+ensemble_scheduling {
+  step [
+    {
+      model_name: "custom_zero_1_float32"
+      model_version: -1
+      input_map {
+        key: "INPUT0"
+        value: "INPUT0"
+      }
+      output_map {
+        key: "OUTPUT0"
+        value: "temp_0"
+      }
+    },
+    {
+      model_name: "libtorch_zero_1_float32"
+      model_version: -1
+      input_map {
+        key: "INPUT__0"
+        value: "temp_0"
+      }
+      output_map {
+        key: "OUTPUT__0"
+        value: "OUTPUT0"
+      }
+    }
+  ]
+}